pyproject.toml
[project]
name = "speekium"
version = "0.1.0"
description = "A smart voice assistant with pluggable LLM backends"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "kanweiwei", email = "kwwnjujlc@gmail.com"}
]
requires-python = ">=3.10"
keywords = [
    "voice-assistant",
    "llm",
    "tts",
    "asr",
    "speech-recognition",
    "text-to-speech",
    "ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "edge-tts",
    "numpy",
    "sounddevice",
    "scipy",
    "torch",
    "torchaudio",
    "funasr",
    "httpx",
    "llvmlite>=0.42.0",
    "pywebview>=6.0",
    "pynput>=1.7.6",  # Global hotkey support
    "pystray>=0.19.5",  # System tray support
    "Pillow>=10.2.0",  # Image support for tray icon
]

[project.optional-dependencies]
gui = ["pywebview"]
piper = ["piper-tts"]  # Offline TTS support

[project.urls]
Homepage = "https://github.com/kanweiwei/speekium"
Repository = "https://github.com/kanweiwei/speekium"
Issues = "https://github.com/kanweiwei/speekium/issues"

[project.scripts]
speekium = "speekium:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
]

[tool.ruff]
# Target Python 3.10+
target-version = "py310"

# Line length
line-length = 100

# Exclude directories
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "node_modules",
    ".pytest_cache",
    "tauri-prototype",
]

[tool.ruff.lint]
# Enable pycodestyle (E), pyflakes (F), and isort (I)
select = ["E", "F", "I", "W", "UP", "B", "SIM", "C90"]

# Ignore specific rules
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function call in argument defaults
    "C901",  # Function too complex (will refactor later)
    "F401",  # Unused imports (will clean up in P2)
    "F841",  # Unused variables (will clean up in P2)
    "F541",  # f-string without placeholders (will clean up in P2)
    "I001",  # Import sorting (isort will handle this)
    "UP035", # Deprecated typing imports (will upgrade in P2)
    "UP045", # Type annotation style (will upgrade in P2)
    "B018",  # Useless expression (will fix in test refactoring)
    "E402",  # Module level import not at top (intentional in tests)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with 4 spaces
indent-style = "space"

# Unix line endings
line-ending = "lf"
